Moved logger.ts, renamed BasicLogger to Logger, and created a main logger "include" file that extension files reference
