+++ src/shared/logger/loggableType.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type Loggable = Error | string
