+++ src/test/shared/logger.test.ts
@@ -7,7 +7,8 @@ import * as assert from 'assert'
 import * as del from 'del'
 import * as path from 'path'
 import * as filesystemUtilities from '../../shared/filesystemUtilities'
-import { createLogger, Logger, OutputChannelTransport, WinstonToolkitLogger } from '../../shared/logger'
+import { Logger } from '../../shared/logger'
+import { createLogger, OutputChannelTransport, WinstonToolkitLogger } from '../../shared/logger/logger'
 import { MockOutputChannel } from '../mockOutputChannel'
 import { assertThrowsError } from './utilities/assertUtils'
